Added a new powerup & Changed time of powerup effects

+++ classes/powerup.cpp
@@ -1,7 +1,7 @@
 // Header
 #include "powerup.h"
 // Constructor
-Powerup::Powerup(const Powerup::Type &Type,const Powerup::Impact &Impact, const int &Chance, const int &Min, const int &Rand){
+Powerup::Powerup(const Powerup::Type &Type,const Powerup::Impact &Impact, const int &Chance, const float &Min, const int &Rand){
     type=Type;
     impact=Impact;
     spawn_chance=Chance;
